updates to dataUseV3
